+++ package.json
@@ -2,7 +2,7 @@
 	"name": "ui5plugin",
 	"displayName": "SAPUI5 Extension",
 	"description": "Extension for working with UI5 projects",
-	"version": "0.5.5",
+	"version": "0.6.0",
 	"publisher": "iljapostnovs",
 	"repository": {
 		"type": "git",
