+++ src/classes/CustomLibMetadata/JSParser/types/AbstractType.ts
