+++ src/classes/CustomLibMetadata/JSParser/types/Array.ts
