+++ src/classes/CustomLibMetadata/JSParser/types/Class.ts
