+++ src/classes/CustomLibMetadata/JSParser/types/Function.ts
