+++ src/classes/CustomLibMetadata/JSParser/types/JSComment.ts
