+++ src/classes/CustomLibMetadata/JSParser/types/JSSwitch.ts
