+++ src/classes/CustomLibMetadata/JSParser/types/Object.ts
