+++ src/classes/CustomLibMetadata/JSParser/types/String.ts
