+++ src/classes/CustomLibMetadata/JSParser/types/TernaryOperation.ts
