+++ src/classes/CustomLibMetadata/JSParser/types/Variable.ts
