+++ src/classes/CustomLibMetadata/UI5Parser/UIClass/AbstractUIClass.ts
