+++ src/classes/StandardLibMetadata/SAPNodeDAO.ts
@@ -1,4 +1,4 @@
-import { SAPNode } from "../StandardLibMetadata/SAPNode";
+import { SAPNode } from "./SAPNode";
 import * as rp from "request-promise";
 import * as vscode from "vscode";
