+++ src/extension.ts
@@ -1,15 +1,8 @@
 import * as vscode from "vscode";
-import { CompletionItemDAO } from "./classes/DAOAndFactories/CompletionItemFactory"
-import { GeneratorFactory } from "./classes/CodeGenerators/GeneratorFactory";
-import { DefineEditor as SAPUIDefineCommand } from "./classes/VSCommands/fileEditors/SAPUIDefineCommand";
-import { ExportToI18NCommand } from "./classes/VSCommands/fileEditors/ExportToI18NCommand";
-import { ClearCacheCommand } from "./classes/VSCommands/ClearCacheCommand";
-import { ViewControllerSwitcher } from "./classes/VSCommands/switchers/ViewControllerSwitcher";
-import { DefineGenerator } from "./classes/CodeGenerators/define/UIDefineCompletionItemGenerator";
-import { WorkspaceCompletionItemDAO } from "./classes/DAOAndFactories/WorkspaceCompletionItemFactory";
-import { UIClassFactory } from "./classes/DAOAndFactories/UIClassFactory";
-import { FileReader } from "./classes/FileReader";
-import { UIClassDefinitionFinder } from "./classes/SyntaxParsers/UI5Parser/UIClass/UIClassDefinitionFinder";
+import { FileReader } from "./classes/Util/FileReader";
+import { CommandRegistrator } from "./classes/Util/CommandRegistrator";
+import { CompletionItemRegistrator } from "./classes/Util/CompletionItemRegistrator";
+import { DefinitionProviderRegistrator } from "./classes/Util/DefinitionProviderRegistrator";
 
 export async function activate(context: vscode.ExtensionContext) {
 	vscode.window.withProgress({
@@ -22,78 +15,13 @@ export async function activate(context: vscode.ExtensionContext) {
 
 		return new Promise(async resolve => {
 			try {
-				/* Commands */
-				let insertUIDefineCommand = vscode.commands.registerCommand("ui5plugin.moveDefineToFunctionParameters", SAPUIDefineCommand.insertUIDefine);
-				context.subscriptions.push(insertUIDefineCommand);
+				CommandRegistrator.register(context);
 
-				let switcherCommand = vscode.commands.registerCommand("ui5plugin.switchBetweenVC", ViewControllerSwitcher.switchBetweenViewController);
-				context.subscriptions.push(switcherCommand);
+				DefinitionProviderRegistrator.register(context);
 
-				let cleacCacheCommand = vscode.commands.registerCommand("ui5plugin.clearCache", ClearCacheCommand.clearCache.bind(undefined, context));
-				context.subscriptions.push(cleacCacheCommand);
+				await CompletionItemRegistrator.register(context, progress);
 
-				let exportToI18NCommand = vscode.commands.registerCommand("ui5plugin.exportToi18n", ExportToI18NCommand.export);
-				context.subscriptions.push(exportToI18NCommand);
-
-				console.log("Commands registered");
-
-				/* Completion Items */
-				let XMLCompletionItemDAO = new CompletionItemDAO(GeneratorFactory.language.xml);
-				let XMLCompletionItems = await XMLCompletionItemDAO.getLanguageSpecificCompletionItems(progress, context);
-				console.log("XML Completion Items generated");
-
-				let JSCompletionItemDAO = new CompletionItemDAO(GeneratorFactory.language.js);
-				let JSDefineCompletionItems = await JSCompletionItemDAO.getLanguageSpecificCompletionItems(progress, context)
-				console.log("JS Completion Items generated");
-
-				let JSMethodPropertyProvider = vscode.languages.registerCompletionItemProvider({language: "javascript", scheme: "file"}, {
-					provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext) {
-						let itemsToReturn:vscode.CompletionItem[] = [];
-						try {
-							if (DefineGenerator.getIfCurrentPositionIsInDefine(position)) {
-								itemsToReturn = JSDefineCompletionItems;
-							}
-
-							itemsToReturn = itemsToReturn.concat(JSCompletionItemDAO.generateUIClassCompletionItems());
-						} catch (error) {
-							console.log(error);
-						}
-						return itemsToReturn;
-					}
-				}, ".");
-
-				let JSViewIDProvider = vscode.languages.registerCompletionItemProvider({language: "javascript", scheme: "file"}, {
-					provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext) {
-						return JSCompletionItemDAO.generateIdCompletionItems();
-					}
-				}, "this.getView().byId(");
-
-				let XMLProvider = vscode.languages.registerCompletionItemProvider({language: "xml", scheme: "file"}, {
-					provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext) {
-						return XMLCompletionItems;
-					}
-				});
-
-				context.subscriptions.push(XMLProvider);
-				context.subscriptions.push(JSMethodPropertyProvider);
-				context.subscriptions.push(JSViewIDProvider);
-
-				/* Events */
-				ClearCacheCommand.subscribeToPropertyChange(context);
-				WorkspaceCompletionItemDAO.subscribeToFileOpening((event: vscode.TextDocument) => {
-					WorkspaceCompletionItemDAO.synchronise(JSDefineCompletionItems, event);
-				});
-
-				UIClassFactory.synchroniseCacheOnDocumentSave();
 				FileReader.synchroniseCacheOnDocumentSave();
-
-				/* Definition provider */
-				const definitionProviderDisposable = vscode.languages.registerDefinitionProvider({language: "javascript", scheme: "file"}, {
-					provideDefinition(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken) {
-						return UIClassDefinitionFinder.getPositionAndUriOfCurrentVariableDefinition();
-					}
-				});
-				context.subscriptions.push(definitionProviderDisposable);
 
 				resolve();
 			} catch (error) {
